10.2_Dodano przykład z zastosowaniem adnotacji @Bean
