+++ kodilla-spring/src/main/java/com/kodilla/spring/SpringRunner.java
@@ -1,5 +1,13 @@
 package com.kodilla.spring;
 
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
 public class SpringRunner {
+
+    public static  void main(String[] args){
+        SpringApplication.run(SpringRunner.class,args);
+    }
 
 }
