+++ kodilla-spring/src/main/java/com/kodilla/spring/shape/Circle.java
@@ -1,0 +1,11 @@
+package com.kodilla.spring.shape;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class Circle implements Shape {
+    @Override
+    public String getShapeName() {
+        return "This is a circle.";
+    }
+}
