+++ kodilla-spring/src/main/java/com/kodilla/spring/shape/Square.java
@@ -1,9 +1,8 @@
 package com.kodilla.spring.shape;
 
-
 public class Square implements Shape {
     @Override
-    public String getShapeName() {
-        return "This is a square.";
+        public String getShapeName() {
+            return "This is a square.";
     }
 }
