+++ kodilla-spring/src/main/java/com/kodilla/spring/shape/Triangle.java
@@ -1,0 +1,11 @@
+package com.kodilla.spring.shape;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class Triangle implements Shape {
+    @Override
+    public String getShapeName() {
+        return "This is a triangle.";
+    }
+}
